create filtering.py and simplify node. Use image.py for I/O.
